M4-pipelines/requirements.txt
apache-airflow==2.9.0
pandas==1.5.3
scikit_learn==1.2.2